Attempt impl of http/2 (for now, abandoned)
https://www.rfc-editor.org/rfc/rfc7540.html is the RFC for http/2
connections, which was revised in
https://www.rfc-editor.org/rfc/rfc9113.html .

The reason this venture was abandoned can be seen in section 3.2.1 of
rfc7540, as it describes a "HTTP2-Settings" HTTP header that is
documented to be "token68" and "base64" in the same document. (token68
is defined here https://www.rfc-editor.org/rfc/rfc7235#section-2.1 ) In
other words, the header must be one or the other, and there is no
distinction for which nor is there any indication of using one over the
other.

For now, this branch will be abandoned.