cxx_impl: Parse ipv6 scope too (like "...%wlan1")
When parsing ipv6 strings, also parse for device scope (which is used
sometimes when using ipv6). This means that a string like "%enp1s0" can
be appended to the ipv6 string and it should properly get the scope id
of the device.

Note that this is very difficult to unit test for as available devices
with ipv6 "fe00::..." addresses vary from system to system.